:bug: fix: the component doesn't show when pressed the start button
