+++ src/components/ButtonBackward.vue
@@ -12,9 +12,9 @@
   color: white;
   font-size: 30px;
   cursor: pointer; 
-  position: absolute;
+  position: fixed;
   display: flex;
   bottom: 10px;
-  right: 30px;
+  right: 9%;
 }
 </style>
