--- src/components/icons/IconPositionMap.vue
@@ -1,7 +1,0 @@
-<template>
-    <svg width="82" height="82" viewBox="0 0 82 82" fill="none" xmlns="http://www.w3.org/2000/svg">
-        <circle cx="40.9807" cy="40.9808" r="30" transform="rotate(60 40.9807 40.9808)" fill="#307DAF"/>
-        <path d="M52.7543 31.8636C54.7534 32.0945 55.9628 34.1892 55.1631 36.0359L48.9067 50.4845C48.8152 50.6958 48.6998 50.8958 48.5626 51.0806L41.4079 60.719C39.7717 62.9232 36.2798 61.9543 36.0132 59.2223L34.768 46.462C34.684 45.6003 34.2317 44.8169 33.5275 44.3133L23.0993 36.8548C20.8666 35.2579 21.7735 31.7494 24.5004 31.4345L36.4249 30.0576C36.6536 30.0312 36.8845 30.0312 37.1132 30.0576L52.7543 31.8636Z" fill="white"/>
-    </svg>
-</template>
-
