+++ src/views/HomeView.vue
@@ -1,7 +1,8 @@
 <script setup lang="ts">
 import BaseMap from "@/components/BaseMap.vue";
+import BaseMapLAS from "@/components/BaseMapLAS.vue";
 </script>
 
 <template>
-    <BaseMap></BaseMap>
+    <BaseMap/>
 </template>
